+++ R/glmmulti.R
@@ -1,25 +1,22 @@
 #' Binomial logistic regression multivariable models: \code{finalfit} model
 #' wrapper
 #'
-#' Using \code{finalfit} conventions, produces multiple multivariable binomial
-#' logistic regression models for a set of explanatory variables against a
+#' Using \code{finalfit} conventions, produces a multivariable binomial
+#' logistic regression model for a set of explanatory variables against a
 #' binary dependent.
 #'
 #' Uses \code{\link[stats]{glm}} with \code{finalfit} modelling conventions.
-#' Output can be passed to \code{\link{fit2df}}. Note that this function can
-#' take multiple \code{dependent} variables as well, but performs multiple
-#' individual models, not a multivariate analysis.
+#' Output can be passed to \code{\link{fit2df}}.
 #'
 #' @param .data Data frame.
-#' @param dependent Character vector usually of length 1, but can take more than
-#'   1 dependent:  name of depdendent variable (must have 2 levels).
+#' @param dependent Character vector of length 1: name of depdendent variable
+#'   (must have 2 levels).
 #' @param explanatory Character vector of any length: name(s) of explanatory
 #'   variables.
 #' @param family Character vector quoted or unquoted of the error distribution
 #'   and link function to be used in the model, see \code{\link[stats]{glm}}.
-#' @param ... Other arguments to pass to \code{\link[stats]{glm}}.  
-#' @return A list of multivariable \code{\link[stats]{glm}} fitted model
-#'   outputs. Output is of class \code{glmlist}.
+#' @param ... Other arguments to pass to \code{\link[stats]{glm}}.
+#' @return A multivariable \code{\link[stats]{glm}} fitted model.
 #'
 #' @seealso \code{\link{fit2df}, \link{finalfit_merge}}
 #' @family finalfit model wrappers
@@ -36,14 +33,8 @@
 #' 	fit2df(estimate_suffix=" (univariable)")
 #' 
 glmmulti <- function(.data, dependent, explanatory, family = "binomial", ...){
-  result = list()
-  for (i in 1:length(dependent)){
-    result[[i]] = ff_eval(
-    	glm(ff_formula(dependent[i], explanatory),
-                      data = .data, family = family, ...)
-    )
-  }
-  result = setNames(result, dependent)
-  class(result) = "glmlist"
-  return(result)
+  result = ff_eval(
+    glm(ff_formula(dependent, explanatory),
+        data = .data, family = family, ...)
+  )
 }